rename main to train
