+++ text_clf/__init__.py
@@ -1,4 +1,4 @@
-from .main import main as text_clf_train
+from .train import main as text_clf_train
 
 __version__ = "0.1.0"
 __all__ = ["text_clf_train"]
