+++ text_clf/__main__.py
@@ -1,4 +1,4 @@
-from .main import main
+from .train import main
 
 if __name__ == "__main__":
     exit(main())
